Moved test to devicelab
